refactor: 重构反CORS请求接口
